[Silver V] Title: 막대기, Time: 0 ms, Memory: 2020 KB -BaekjoonHub

+++ 백준/Silver/1094. 막대기/막대기.cc
@@ -1,3 +1,3 @@
-#include <bits/stdc++.h>
+#include<bits/stdc++.h>
 using namespace std;
-int main(){unsigned long long x;cin>>x;cout<<popcount(x);}
+int main(){unsigned x;cin>>x;cout<<popcount(x);}
